requirements.txt
python-telegram-bot==21.9
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.1.0
requests==2.31.0
aiohttp==3.9.5